generic schema added
